i did some stuff!!1!!111
Added a nav bar instead of div, i also changed the font because the previous one was not monospaced.
I could not manage to make the other two words "glitchable" without adding separate strings of Js for them which i will be trying to solve another time.  There is now also a hover effect on the words that inverts the colors of text and background