YES ASIAJSFOIUFIAUHF ITS WOOORKIIING
I worked with the rotate potato SVGS to the left of the text. i had problems with them not rotating from the centre point of themselves when i used "position: absolute" and "transform: translate" but now i managed to use the "float" property on the text and SVG to make them in line with each other. I had a lot of trouble making the smaller svg work (i almost ripped all my hair out) since it had a translate property inherit in it's svg code so it could be centred inside the larger circle. Diana said "two circles" which made me realize that if i duplicate the larger circle and scale it down using the "scale" modifier i can have two circles that are smaller and bigger with the same transform point. I just had to divide 6 (the original line width) with 0.7 to get the new one for the smaller circle since the "scale" property also made the line thinner