I ADDED SO MUCH STUFF GRAAAAAAAHHH
I added some cards for the projects page, i am very proud of the formatting of my code right now and how fast the workflow went for making the cards.
I imported the svg for the icon and THE CODE IS SO LONG, which is to be expected.  I imported the Cherry swash font from google fonts because we use it for the DnD page. I am yet to format the text but yes i am happy about what i did today