Mr. Spumpen
Today i made the lines work as they should. For some reason the box flow was all messed up whenever i placed the lines after the text and header in the HTML but if i placed them first in the div it worked as intended with "flow: right". So at the moment it is just placed with a "margin-top" and i decreased the margins on the text and header margins to account for the lines.

I also made a PLACEHOLDER IMAGE WOOO 🥳. It is my character Mr. Spumpen that i drew in aseprite. I made it to go with the Lipsum text on the cards that i havent used yet. it was mostly to make sure thatt he universal settings worked as intended.